backend/requirements.txt
fastapi
python-dotenv
google-generativeai
supabase
jwt
fastapi-sessions
itsdangerous